% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{base_url}
\alias{base_url}
\title{Get the base URL for the ServCat API}
\usage{
base_url(secure = FALSE)
}
\arguments{
\item{secure}{Logical. Use the secure ServCat API?}
}
\value{
A length-1 character vector
}
\description{
Get the base URL for the ServCat API
}
\keyword{internal}
