% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{chunk_reference_ids}
\alias{chunk_reference_ids}
\title{Chunk reference IDs for Profile requests}
\usage{
chunk_reference_ids(reference_ids, chunk_size = 25)
}
\arguments{
\item{reference_ids}{Numeric vector of reference IDs}

\item{chunk_size}{Maximum number of IDs per request}
}
\value{
A list of numeric vectors
}
\description{
Chunk reference IDs for Profile requests
}
\keyword{internal}
