% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{collapse_reference_ids}
\alias{collapse_reference_ids}
\title{Collapse reference IDs for the Profile endpoint}
\usage{
collapse_reference_ids(reference_ids)
}
\arguments{
\item{reference_ids}{Numeric vector of reference IDs}
}
\value{
A length-1 character vector
}
\description{
Collapse reference IDs for the Profile endpoint
}
\keyword{internal}
