% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lists.R
\name{get_access_constraints}
\alias{get_access_constraints}
\title{Get valid ServCat access constraint text}
\usage{
get_access_constraints(secure = FALSE, api_key = NULL)
}
\arguments{
\item{secure}{Logical. Use the secure API?}

\item{api_key}{Optional secure API key.}
}
\value{
A tibble with access constraint metadata.
}
\description{
Retrieves the fixed list of access constraint text used when setting
permissions and constraints for a ServCat reference.
}
\keyword{internal}
