% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_api_key}
\alias{get_api_key}
\title{Retrieve secure ServCat API key}
\usage{
get_api_key(env_var = "SERVCAT_API_KEY")
}
\arguments{
\item{env_var}{Environment variable name.}
}
\value{
A length-1 character vector.
}
\description{
Retrieve secure ServCat API key
}
\keyword{internal}
