% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lists.R
\name{get_date_precisions}
\alias{get_date_precisions}
\title{Get valid ServCat date precision choices}
\usage{
get_date_precisions(secure = FALSE, api_key = NULL)
}
\arguments{
\item{secure}{Logical. Use the secure API?}

\item{api_key}{Optional secure API key.}
}
\value{
A tibble with date precision choices.
}
\description{
Retrieves the fixed list of DatePrecision choices used for issued dates and
content begin/end dates.
}
\keyword{internal}
