% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lists.R
\name{get_file_processing_tags}
\alias{get_file_processing_tags}
\title{Get valid ServCat file processing tags}
\usage{
get_file_processing_tags(secure = FALSE, api_key = NULL)
}
\arguments{
\item{secure}{Logical. Use the secure API?}

\item{api_key}{Optional secure API key.}
}
\value{
A tibble with file processing tag choices.
}
\description{
Retrieves the fixed list of tags used when designating certain digital file
types for offline processing.
}
\keyword{internal}
