% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_fixed_list}
\alias{get_fixed_list}
\title{Get a ServCat fixed list}
\usage{
get_fixed_list(endpoint, secure = FALSE, api_key = NULL, rename_key = TRUE)
}
\arguments{
\item{endpoint}{Fixed-list endpoint name.}

\item{secure}{Logical. Use the secure API?}

\item{api_key}{Optional secure API key.}

\item{rename_key}{Logical. Rename a \code{key} column to \code{code}?}
}
\value{
A tibble.
}
\description{
Get a ServCat fixed list
}
\keyword{internal}
