% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_lifecycle_info}
\alias{get_lifecycle_info}
\title{Get lifecycle information from the secure ServCat API}
\usage{
get_lifecycle_info(reference_id, api_key = NULL)
}
\arguments{
\item{reference_id}{A single ServCat reference ID.}

\item{api_key}{Optional secure API key. If omitted, the package API-key
helper is used.}
}
\value{
A named list.
}
\description{
Get lifecycle information from the secure ServCat API
}
\examples{
\dontrun{
# Retrieve lifecycle information from the secure API
get_lifecycle_info(140411)
}

}
