% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lists.R
\name{get_publisher_display_name}
\alias{get_publisher_display_name}
\title{Get ServCat publisher display name by reference type}
\usage{
get_publisher_display_name(reference_type, secure = FALSE, api_key = NULL)
}
\arguments{
\item{reference_type}{A single ServCat reference type, such as \code{"Brochure"}.}

\item{secure}{Logical. Use the secure API?}

\item{api_key}{Optional secure API key.}
}
\value{
A tibble with the publisher field display name.
}
\description{
Retrieves the display name for the \code{Publisher} field for a given ServCat
reference type.
}
\keyword{internal}
