% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lists.R
\name{get_reference_type_groups}
\alias{get_reference_type_groups}
\title{Get valid ServCat reference type groups}
\usage{
get_reference_type_groups(secure = FALSE, api_key = NULL)
}
\arguments{
\item{secure}{Logical. Use the secure API?}

\item{api_key}{Optional secure API key.}
}
\value{
A tibble with reference type group choices.
}
\description{
Retrieves the fixed list of Reference Type Groups used as an Advanced Search
option in ServCat.
}
\keyword{internal}
