% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lists.R
\name{get_reference_types}
\alias{get_reference_types}
\title{Get valid ServCat reference types}
\usage{
get_reference_types(secure = FALSE, api_key = NULL)
}
\arguments{
\item{secure}{Logical. Use the secure API?}

\item{api_key}{Optional secure API key}
}
\value{
A tibble with ServCat reference type metadata
}
\description{
Get valid ServCat reference types
}
\keyword{internal}
