% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lists.R
\name{get_service_version}
\alias{get_service_version}
\title{Get the current ServCat application version}
\usage{
get_service_version(secure = FALSE, api_key = NULL)
}
\arguments{
\item{secure}{Logical. Use the secure API?}

\item{api_key}{Optional secure API key.}
}
\value{
A length-1 character vector with the current ServCat application
version.
}
\description{
Retrieves the current ServCat application version from the service metadata
endpoint.
}
\keyword{internal}
