% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lists.R
\name{get_subject_categories}
\alias{get_subject_categories}
\title{Get valid ServCat subject categories}
\usage{
get_subject_categories(secure = FALSE, api_key = NULL)
}
\arguments{
\item{secure}{Logical. Use the secure API?}

\item{api_key}{Optional secure API key.}
}
\value{
A tibble with subject category choices.
}
\description{
Retrieves the fixed list of Subject Categories used on the Keywords and
Category tab in ServCat.
}
\keyword{internal}
