% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lists.R
\name{get_web_service_types}
\alias{get_web_service_types}
\title{Get valid ServCat web service types}
\usage{
get_web_service_types(secure = FALSE, api_key = NULL)
}
\arguments{
\item{secure}{Logical. Use the secure API?}

\item{api_key}{Optional secure API key.}
}
\value{
A tibble with web service type choices.
}
\description{
Retrieves the fixed list of Web Service, or Linked Resource, types used by
ServCat.
}
\keyword{internal}
