% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lists.R
\name{list_file_tags}
\alias{list_file_tags}
\title{List valid ServCat file processing tags}
\usage{
list_file_tags(secure = FALSE, api_key = NULL)
}
\arguments{
\item{secure}{Logical. Use the secure API?}

\item{api_key}{Optional secure API key. If omitted, the package API-key
helper is used for secure requests.}
}
\value{
A tibble with file processing tag choices.
}
\description{
Retrieves the fixed list of tags used when designating certain digital file
types for offline processing.
}
\examples{
\dontrun{
# List public file processing tags
list_file_tags()

# List file processing tags from the secure API
list_file_tags(secure = TRUE)
}

}
