% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lists.R
\name{list_subject_categories}
\alias{list_subject_categories}
\title{List valid ServCat subject categories}
\usage{
list_subject_categories(secure = FALSE, api_key = NULL)
}
\arguments{
\item{secure}{Logical. Use the secure API?}

\item{api_key}{Optional secure API key. If omitted, the package API-key
helper is used for secure requests.}
}
\value{
A tibble with subject category choices.
}
\description{
Retrieves the fixed list of Subject Categories used on the Keywords and
Category tab in ServCat.
}
\examples{
\dontrun{
# List public subject category choices
list_subject_categories()

# List subject category choices from the secure API
list_subject_categories(secure = TRUE)
}

}
