% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lists.R
\name{list_web_service_types}
\alias{list_web_service_types}
\title{List valid ServCat web service types}
\usage{
list_web_service_types(secure = FALSE, api_key = NULL)
}
\arguments{
\item{secure}{Logical. Use the secure API?}

\item{api_key}{Optional secure API key. If omitted, the package API-key
helper is used for secure requests.}
}
\value{
A tibble with web service type choices.
}
\description{
Retrieves the fixed list of Web Service, or Linked Resource, types used by
ServCat.
}
\examples{
\dontrun{
# List public web service type choices
list_web_service_types()

# List web service type choices from the secure API
list_web_service_types(secure = TRUE)
}

}
