% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lists_to_tibbles}
\alias{lists_to_tibbles}
\title{Convert nested lists in a profile to tibbles}
\usage{
lists_to_tibbles(parent_list, child_list_names)
}
\arguments{
\item{parent_list}{A ServCat profile record.}

\item{child_list_names}{Element names to convert.}
}
\value{
A modified list.
}
\description{
Convert nested lists in a profile to tibbles
}
\keyword{internal}
