% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lists_to_vectors}
\alias{lists_to_vectors}
\title{Convert nested lists in a profile to vectors}
\usage{
lists_to_vectors(parent_list, child_list_names)
}
\arguments{
\item{parent_list}{A ServCat profile record}

\item{child_list_names}{Element names to simplify}
}
\value{
A modified list
}
\description{
Convert nested lists in a profile to vectors
}
\keyword{internal}
