% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{profile_url}
\alias{profile_url}
\title{Get a ServCat profile URL}
\usage{
profile_url(reference_id)
}
\arguments{
\item{reference_id}{A ServCat reference ID.}
}
\value{
A length-1 character vector
}
\description{
Get a ServCat profile URL
}
\keyword{internal}
