% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sanitize_file_name}
\alias{sanitize_file_name}
\title{Sanitize a file name from ServCat metadata}
\usage{
sanitize_file_name(file_name, fallback)
}
\arguments{
\item{file_name}{File name from the ServCat API.}

\item{fallback}{Fallback file name.}
}
\value{
A length-1 character vector.
}
\description{
Sanitize a file name from ServCat metadata
}
\keyword{internal}
