% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{search_references_by_id}
\alias{search_references_by_id}
\title{Search ServCat references by ID}
\usage{
search_references_by_id(reference_ids, secure = FALSE, api_key = NULL)
}
\arguments{
\item{reference_ids}{Numeric vector of ServCat reference IDs.}

\item{secure}{Logical. Use the secure API?}

\item{api_key}{Optional secure API key. If omitted, the package API-key
helper is used for secure requests.}
}
\value{
A named list containing detailed information about each reference.
}
\description{
Search ServCat references by ID
}
\examples{
\dontrun{
# Retrieve detailed information for one reference
search_references_by_id(140411)

# Retrieve detailed information for multiple references
search_references_by_id(c(140411, 140412))

# Use the secure API
search_references_by_id(
  reference_ids = 140411,
  secure = TRUE
)
}

}
