% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validate_flag}
\alias{validate_flag}
\title{Validate a TRUE/FALSE argument}
\usage{
validate_flag(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{Value to validate}

\item{arg}{Used to generate helpful error messages}

\item{call}{Calling environment}
}
\value{
\code{NULL}, invisibly.
}
\description{
Validate a TRUE/FALSE argument
}
\keyword{internal}
