% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validate_response}
\alias{validate_response}
\title{Validate a ServCat API response}
\usage{
validate_response(resp, nice_msg_400, nice_msg_500, call = rlang::caller_env())
}
\arguments{
\item{resp}{An \code{httr2_response}}

\item{nice_msg_400}{Optional message for 4xx errors}

\item{nice_msg_500}{Optional message for 5xx errors}

\item{call}{Calling environment}
}
\value{
\code{NULL}, invisibly.
}
\description{
Validate a ServCat API response
}
\keyword{internal}
