% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validate_string}
\alias{validate_string}
\title{Validate a single non-empty character argument}
\usage{
validate_string(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{Value to validate.}

\item{arg}{Used to generate helpful error messages.}

\item{call}{Calling environment.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Validate a single non-empty character argument
}
\keyword{internal}
