% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validate_whole_number}
\alias{validate_whole_number}
\title{Validate a positive whole-number argument}
\usage{
validate_whole_number(
  x,
  arg = rlang::caller_arg(x),
  multiple_ok = FALSE,
  min = 1,
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{Value to validate.}

\item{arg}{Used to generate helpful error messages.}

\item{multiple_ok}{Logical. Are multiple values allowed?}

\item{min}{Minimum allowed value.}

\item{call}{Calling environment.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Validate a positive whole-number argument
}
\keyword{internal}
