% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validate_reference_id}
\alias{validate_reference_id}
\title{Validate a ServCat reference ID}
\usage{
validate_reference_id(
  reference_id,
  multiple_ok = FALSE,
  arg = rlang::caller_arg(reference_id),
  call = rlang::caller_env()
)
}
\arguments{
\item{reference_id}{ServCat reference ID or IDs.}

\item{multiple_ok}{Logical. Are multiple IDs allowed?}

\item{arg}{Used to generate helpful error messages.}

\item{call}{Calling environment.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Validate a ServCat reference ID
}
\keyword{internal}
